+++ test/resources/attachment-constraints-doc-definitions.js
@@ -8,14 +8,16 @@
       maximumTotalSize: 40,
       maximumAttachmentCount: 3,
       supportedExtensions: [ 'html', 'jpg', 'pdf', 'txt', 'xml' ],
-      supportedContentTypes: [ 'text/html', 'image/jpeg', 'application/pdf', 'text/plain', 'application/xml' ]
+      supportedContentTypes: [ 'text/html', 'image/jpeg', 'application/pdf', 'text/plain', 'application/xml' ],
+      filenameRegexPattern: /^(foo|ba[rz]|qux)\.[a-z]+$/
     },
     propertyValidators: {
       attachmentRefProp: {
         type: 'attachmentReference',
         maximumSize: 40,
         supportedExtensions: [ 'foo', 'html', 'jpg', 'pdf', 'txt', 'xml' ],
-        supportedContentTypes: [ 'text/bar', 'text/html', 'image/jpeg', 'application/pdf', 'text/plain', 'application/xml' ]
+        supportedContentTypes: [ 'text/bar', 'text/html', 'image/jpeg', 'application/pdf', 'text/plain', 'application/xml' ],
+        regexPattern: /^[a-z]+\.[a-z]+$/
       }
     }
   },